requirements.txt
streamlit==1.39.0
pandas
numpy
faiss-cpu
scikit-learn
sentence-transformers
torch
fastapi
uvicorn
python-dotenv
joblib
